Removendo os alunos em modo offline

+++ app/src/main/java/com/pdxdev/agenda/dao/AlunoDAO.java
@@ -20,7 +20,7 @@ import java.util.UUID;
 public class AlunoDAO extends SQLiteOpenHelper{
 
     public AlunoDAO(Context context) {
-        super(context, "Agenda", null, 5);
+        super(context, "Agenda", null, 6);
     }
 
     @Override
@@ -32,7 +32,8 @@ public class AlunoDAO extends SQLiteOpenHelper{
                 "site TEXT, " +
                 "nota REAL, " +
                 "caminhoFoto TEXT," +
-                "sincronizado INT DEFAULT 0);";
+                "sincronizado INT DEFAULT 0," +
+                "desativado INT DEFAULT 0);";
         db.execSQL(sql);
     }
 
@@ -80,7 +81,9 @@ public class AlunoDAO extends SQLiteOpenHelper{
             case 4:
                 String adicionaCampoSincronizado = "ALTER TABLE Alunos ADD COLUMN sincronizado INT DEFAULT 0";
                 db.execSQL(adicionaCampoSincronizado);
-
+            case 5:
+                String adicionaCampoDesativado = "ALTER TABLE Alunos ADD COLUMN desativado INT DEFAULT 0";
+                db.execSQL(adicionaCampoDesativado);
         }
 
     }
@@ -90,7 +93,7 @@ public class AlunoDAO extends SQLiteOpenHelper{
     }
 
     public List<Aluno> buscaAlunos() {
-        String sql = "SELECT * FROM Alunos;";
+        String sql = "SELECT * FROM Alunos WHERE desativado = 0;";
         SQLiteDatabase db = getReadableDatabase();
         Cursor c = db.rawQuery(sql, null);
 
@@ -113,6 +116,7 @@ public class AlunoDAO extends SQLiteOpenHelper{
             aluno.setNota(c.getDouble(c.getColumnIndex("nota")));
             aluno.setCaminhoFoto(c.getString(c.getColumnIndex("caminhoFoto")));
             aluno.setSincronizado(c.getInt(c.getColumnIndex("sincronizado")));
+            aluno.setDesativado(c.getInt(c.getColumnIndex("desativado")));
 
             alunos.add(aluno);
         }
@@ -145,8 +149,12 @@ public class AlunoDAO extends SQLiteOpenHelper{
         SQLiteDatabase db = getWritableDatabase();
 
         String [] params = {aluno.getId()};
-        db.delete("Alunos", "id = ?", params);
-
+        if(aluno.estaDesativado()){
+            db.delete("Alunos", "id = ?", params);
+        } else {
+            aluno.desativa();
+            altera(aluno);
+        }
     }
 
     @NonNull
@@ -160,6 +168,7 @@ public class AlunoDAO extends SQLiteOpenHelper{
         dados.put("nota", aluno.getNota());
         dados.put("caminhoFoto", aluno.getCaminhoFoto());
         dados.put("sincronizado", aluno.getSincronizado());
+        dados.put("desativado", aluno.getDesativado());
         return dados;
     }
